+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Alle noemenswaardige wijzigingen aan dit project worden in dit bestand bijgehoud
 Het formaat is gebaseerd op [Keep a Changelog](https://keepachangelog.com/nl/1.1.0/)
 en dit project volgt [Semantic Versioning](https://semver.org/lang/nl/).
 
-## [Unreleased]
+## [1.8.0] - 2026-09-14
 
 ### Toegevoegd
 - CI controleert of de broncode met swift-format is opgemaakt en faalt als dat niet zo is